whyml-cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whyml-cli"
version = "0.1.0"
description = "WhyML CLI - Unified command-line interface for WhyML ecosystem"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Tom Sapletta", email = "tom@sapletta.pl"}
]
maintainers = [
    {name = "Tom Sapletta", email = "tom@sapletta.pl"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Text Processing :: Markup",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: System :: Systems Administration",
    "Environment :: Console",
]
keywords = [
    "whyml", "cli", "command-line", "scraping", "conversion", "validation",
    "code-generation", "manifest", "html", "react", "vue", "php"
]
requires-python = ">=3.8"
dependencies = [
    "whyml-core>=0.1.0",
    "aiohttp>=3.8.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
# Full ecosystem with all functionality
full = [
    "whyml-scrapers>=0.1.0",
    "whyml-converters>=0.1.0",
]
# Web scraping functionality
scraping = [
    "whyml-scrapers>=0.1.0",
]
# Code generation functionality
conversion = [
    "whyml-converters>=0.1.0",
]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
# Testing dependencies
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-click>=1.1.0",
]
# Documentation dependencies
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=1.0.0",
]

[project.scripts]
whyml-cli = "whyml_cli.main:cli_main"
whyml = "whyml_cli.main:cli_main"

[project.urls]
Homepage = "https://github.com/dynapsys/whyml"
Documentation = "https://github.com/dynapsys/whyml/docs"
Repository = "https://github.com/dynapsys/whyml"
"Bug Tracker" = "https://github.com/dynapsys/whyml/issues"
Changelog = "https://github.com/dynapsys/whyml/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["whyml_cli*"]

[tool.setuptools.package-dir]
"" = "."

# Black configuration
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["whyml_cli", "tests"]

# MyPy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with -m 'not slow')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "cli: marks tests as CLI tests",
]
asyncio_mode = "auto"

# Coverage configuration
[tool.coverage.run]
source = ["whyml_cli"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self\\.debug:",
    "if settings\\.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"
